requirements.txt
streamlit
pandas
jinja2
geopy
pyswisseph
st-gsheets-connection